CORRECIONES

+++ Reg_Merca_WebApp/Inicio/recuperar.aspx.vb
@@ -27,6 +27,7 @@
     End Sub
 
     Private Sub bttContinuar_Click(sender As Object, e As EventArgs) Handles bttContinuar.Click
+        'comprobar que el usuario existe y que este activo o bloqueado para permitir ingreso
         Dim Ssql As String = String.Empty
         Ssql = "SELECT  * FROM DB_Nac_Merca.tbl_02_usuarios   where usuario = BINARY  '" & txtUsuarioPreguntas.Text & "';"
         Using con As New ControlDB
@@ -35,13 +36,12 @@
         End Using
         If Session("NumReg") > 0 Then
             Dim registro As DataRow = DataSetX.Tables(0).Rows(0)
-            If CInt(registro("estado")) = 0 Or CInt(registro("estado")) = 3 Or CInt(registro("estado")) = 5 Then
+            If CInt(registro("estado")) <> 2 And CInt(registro("estado")) <> 4 Then
                 Page.ClientScript.RegisterStartupScript(Me.GetType(), "alert", "<script type=""text/javascript"">swal('Advertencia','Este usuario puede estar inactivo, caducado o sin completar el registro.', 'warning');</script>")
             Else
-                Session("usuarioPreguntas") = registro("id_usuario")
+                Session("id_usuarioPreguntas") = registro("id_usuario")
                 Response.Redirect("~/Inicio/preguntas.aspx")
             End If
-            'Page.ClientScript.RegisterStartupScript(Me.GetType(), "alert", "<script type=""text/javascript"">swal('EXITO','Usuario  encontrado', 'success');</script>")
         Else
             Page.ClientScript.RegisterStartupScript(Me.GetType(), "alert", "<script type=""text/javascript"">swal('Advertencia','Usuario no encontrado', 'warning');</script>")
         End If
@@ -55,32 +55,34 @@
             Session("NumReg") = DataSetX.Tables(0).Rows.Count
         End Using
         If Session("NumReg") > 0 Then
-            'enviar correo electronico con token de nueva contraseña
-            Dim registro As DataRow
-            registro = DataSetX.Tables(0).Rows(0)
-            Dim activationCode As String = Guid.NewGuid().ToString()
+            Dim registro As DataRow = DataSetX.Tables(0).Rows(0)
 
-            Ssql = "DELETE FROM `DB_Nac_Merca`.`tbl_35_activacion_usuario` WHERE id_usuario=" & registro("id_usuario") & ";"
-            Using con As New ControlDB
-                con.GME(Ssql, ControlDB.TipoConexion.Cx_Aduana)
-            End Using
+            'solo los usuarios activos
+            If CInt(registro("estado")) <> 2 Then
+                Page.ClientScript.RegisterStartupScript(Me.GetType(), "alert", "<script type=""text/javascript"">swal('Advertencia','Este usuario puede estar bloqueado, inactivo, caducado o sin completar el registro.', 'warning');</script>")
+            Else
+                'enviar correo electronico con token de nueva contraseña
+                Dim activationCode As String = Guid.NewGuid().ToString()
+                Ssql = "DELETE FROM `DB_Nac_Merca`.`tbl_35_activacion_usuario` WHERE id_usuario=" & registro("id_usuario") & ";"
+                Using con As New ControlDB
+                    con.GME(Ssql, ControlDB.TipoConexion.Cx_Aduana)
+                End Using
 
-            Ssql = "INSERT INTO `DB_Nac_Merca`.`tbl_35_activacion_usuario` (`id_usuario`, `codigo_activacion`, `vencimiento`,`tipo`,`estado`) VALUES (" & registro("id_usuario") & ", '" & activationCode & "',DATE_ADD(CONVERT_TZ(NOW(), @@session.time_zone, '-6:00'), INTERVAL " & Application("ParametrosSYS")(8) & " DAY),'clave',1);"
-            Using con As New ControlDB
-                con.GME(Ssql, ControlDB.TipoConexion.Cx_Aduana)
-            End Using
+                Ssql = "INSERT INTO `DB_Nac_Merca`.`tbl_35_activacion_usuario` (`id_usuario`, `codigo_activacion`, `vencimiento`,`tipo`,`estado`) VALUES (" & registro("id_usuario") & ", '" & activationCode & "',DATE_ADD(CONVERT_TZ(NOW(), @@session.time_zone, '-6:00'), INTERVAL " & Application("ParametrosSYS")(8) & " DAY),'clave',1);"
+                Using con As New ControlDB
+                    con.GME(Ssql, ControlDB.TipoConexion.Cx_Aduana)
+                End Using
 
-            Using xCorreo As New ControlCorreo
-                xCorreo.envio_correo("Recibimos una solicitud para restablecer su contraseña. Si no realizó esta solicitud, entonces ignora esta notificación. De lo contrario, puedes restablecer la contraseña mediante este enlace. ", "RESTABLECER CONTRASEÑA",
-                                     txtEmail.Text, Application("ParametrosADMIN")(9), Application("ParametrosADMIN")(11),
-                                     registro("nombre"),
-                                     Request.Url.AbsoluteUri.Replace("recuperar", Convert.ToString("activacion.aspx?ActivationCode=") & activationCode),
-                                     "Restablecer Contraseña",
-                                     Application("ParametrosADMIN")(15), Application("ParametrosADMIN")(10))
-            End Using
-            Response.Redirect("~/Inicio/login.aspx?acction=newsolicitud")
-
-
+                Using xCorreo As New ControlCorreo
+                    xCorreo.envio_correo("Recibimos una solicitud para restablecer su contraseña. Si no realizó esta solicitud, ignore esta notificación. De lo contrario, puede restablecer su contraseña mediante este enlace. ", "RESTABLECER CONTRASEÑA",
+                                         txtEmail.Text, Application("ParametrosADMIN")(9), Application("ParametrosADMIN")(11),
+                                         registro("nombre"),
+                                         Request.Url.AbsoluteUri.Replace("recuperar", Convert.ToString("activacion.aspx?ActivationCode=") & activationCode),
+                                         "Restablecer Contraseña",
+                                         Application("ParametrosADMIN")(15), Application("ParametrosADMIN")(10))
+                End Using
+                Response.Redirect("~/Inicio/login.aspx?acction=newsolicitud")
+            End If
         Else
             Page.ClientScript.RegisterStartupScript(Me.GetType(), "alert", "<script type=""text/javascript"">swal('Correo Electronico','Correo no enviado', 'error');</script>")
         End If
